requirements.txt
discord.py>=2.3.2,<3
aiohttp>=3.9